Added fuel crud

+++ database/migrations/vehicle/2024_02_07_182920_create_vehicle_fuels_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -19,9 +18,9 @@ return new class extends Migration
             $table->string('fuel_station_name')->nullable();
             $table->string('fuel_station_id')->nullable();
             $table->string('fuel_type')->nullable();
-            $table->integer('fuel_quantity');
-            $table->integer('fuel_price')->nullable();
-            $table->integer('total_price')->nullable();
+            $table->decimal('fuel_quantity', 8, 2);
+            $table->decimal('fuel_price', 8, 2)->nullable();
+            $table->decimal('total_price', 8, 2)->nullable();
             $table->text('notes')->nullable();
             $table->foreignId('created_by_id')->nullable()->constrained('users');
             $table->foreignId('last_edited_by_id')->nullable()->constrained('users');
